fix policies bugs

+++ app/controllers/cruds_controller.ts
@@ -1,52 +1,71 @@
 import type { HttpContext } from '@adonisjs/core/http'
 
 export default class CrudsController {
-  public model: any; 
+  public model: any;
+  public policy: any;
 
   //display all data
-  public async index({ response }: HttpContext) {
-    const data = await this.model.all()
+  public async index({ response, request, auth}: HttpContext) {
+    let params = request.qs();
 
-    return response.status(200).json(data);
+    const result = await this.model.listOptions(params, auth.user!.id)
+
+    return response.status(200).json(result);
   }
 
   //display single data by id
-  public async show({ response, params }: HttpContext) {
-    const data = await this.model.findOrFail(params.id);
+  public async show({ response, params, bouncer }: HttpContext) {
+    const data = await this.model.firstOrFailWithPreloads(params.id);
+
+    if (await bouncer.with(this.policy).denies('show', data)) {
+      return response.forbidden('Cannot see this post')
+    }
 
     return response.status(200).json(data);
   }
 
   //store data
-  public async store({ request, response }: HttpContext) {
+  public async store({ request, response, auth }: HttpContext) {
     const payloud = await request.validateUsing(this.model.validator);
-    const result = await this.model.create(payloud);
+    const result = await this.model.create({ ...payloud, user_id: auth.user!.id });
 
     return response.status(201).json(result);
   }
 
   //update single data
-  public async update({ request, response, params }: HttpContext) {
+  public async update({ request, response, params, bouncer }: HttpContext) {
     const data = await this.model.findOrFail(params.id);
-
     const payload = await request.validateUsing(this.model.validator);
 
-    
+    if (await bouncer.with(this.policy).denies('edit', data)) {
+      return response.forbidden('Cannot edit this post')
+    }
+
     data.merge(payload);
+
     await data.save();
 
     return response.status(204).json(data);
   }
 
   //delete single data
-  public async destroy({ response, params }: HttpContext) {
+  public async destroy({ response, params, bouncer }: HttpContext) {
     const data = await this.model.findOrFail(params.id);
+
+    if(await bouncer.with(this.policy).denies('delete', data)){
+      return response.forbidden('Cannot delete this post')
+    }
+
     data.delete();
     return response.status(202);
   }
 
-  public async softDelete({ response, params }: HttpContext) {
+  public async softDelete({ response, params, bouncer }: HttpContext) {
     const data = await this.model.findOrFail(params.id);
+
+    if (await bouncer.with(this.policy).denies('edit', data)) {
+      return response.forbidden('Cannot delete this post')
+    }
 
     data.is_deleted = true;
 
@@ -55,10 +74,15 @@ export default class CrudsController {
     return response.status(204);
   }
 
-  public async recovery({ response, params }: HttpContext) {
+  public async recovery({ response, params, bouncer }: HttpContext) {
     const data = await this.model.findOrFail(params.id);
 
+    if (await bouncer.with(this.policy).denies('edit', data)) {
+      return response.forbidden('Cannot edit this post')
+    }
+
     data.is_deleted = false;
+
     await data.save();
 
     return response.status(204);
